refactor: consolidate Gemini generation utilities into a new module and update import paths.
