+++ integrations/gemini/actions/response/pipeline/03_segmentOutput/main.ts
@@ -12,9 +12,11 @@
 import { log } from "@bott/common";
 import type { ShallowBottEvent } from "@bott/system";
 import { BottEventType } from "@bott/system";
-import { generateFromEvents } from "../../../../generate/fromEvents/main.ts";
-import { getEventSkeletonSchema } from "../../../../generate/fromEvents/skeleton/schema.ts";
-import { skeletonToShallowEvent } from "../../../../generate/fromEvents/skeleton/shallowEvent.ts";
+import {
+  generateFromEvents,
+  getEventSkeletonSchema,
+  skeletonToShallowEvent,
+} from "../../../../generate/module.ts";
 import type { EventPipelineProcessor } from "../types.ts";
 
 const systemPrompt = await Deno.readTextFile(
