+++ integrations/gemini/generate/TODO
@@ -1,4 +1,3 @@
-- [ ] module.ts?
 - [ ] song generation
 - [ ] movie generation
 - [ ] video generation
