+++ integrations/gemini/generate/module.ts
@@ -1,0 +1,15 @@
+/**
+ * @license
+ * This file is part of Bott.
+ *
+ * This project is dual-licensed:
+ * - Non-commercial use: AGPLv3 (see LICENSE file for full text).
+ * - Commercial use: Proprietary License (contact D@nielLaCos.se for details).
+ *
+ * Copyright (C) 2025 DanielLaCos.se
+ */
+
+export * from "./fromEvents/main.ts";
+export * from "./fromEvents/skeleton/schema.ts";
+export * from "./fromEvents/skeleton/shallowEvent.ts";
+export * from "./fromEvents/types.ts";
